old-tests/submission/Microsoft/forms/placeholder.html

<!DOCTYPE html>
<html><head>
        <title>placeholder on input element</title>
        <meta content="text/html; charset=UTF-8" http-equiv="content-type">
        <meta content="placeholder atrribute on elements" name="description">
    <link href="http://www.w3.org/TR/html5/the-input-element.html#dom-input-placeholder" rel="help">
</head>
    <body>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>

        <div id="log"></div>
    
        <div style="display : none">
        <input placeholder="placeholder text" type="text">
        <input placeholder="" type="text">
        <textarea placeholder="placeholder text" type="text"></textarea>
        </div>
        
        <script type="text/javascript">
        
        test(function() {assert_equals(document.getElementsByTagName("input")[0].placeholder, "placeholder text")}, "placeholder support on input element");
        test(function() {assert_equals(document.getElementsByTagName("input")[1].placeholder, "")}, "default value for placeholder on input element");
        test(function() {assert_equals(document.getElementsByTagName("input")[1].getAttribute("placeholder"), "")}, "default value for placeholder when retrieved using getAttribute");
        test(function() {assert_equals(document.getElementsByTagName("textarea")[0].placeholder, "placeholder text")}, "placeholder support on textarea element");
        
        </script>
    
</body></html>
